Refactoring and simplificatoin:
- Tiny refactoring and simplificaiton of abstracts
- refactoring of form field reordering with custom hook
