+++ src/style/abstract/_variables.scss
@@ -14,6 +14,7 @@ $section-max-width: 50%;
 $standard-gap: pxToRem(20);
 $gap-large: pxToRem(40);
 $gap-small: pxToRem(10);
+$gap-xsmall: pxToRem(5);
 $standard-padding: pxToRem(20);
 
 // Form elements
