+++ src/style/pages/tokenCreation/_tokenPreview.scss
@@ -37,7 +37,7 @@
     }
 
     input, select {
-      @extend %form-field;
+      @extend %input-styles;
     }
   }
